+++ gatsby-browser.tsx
@@ -3,11 +3,11 @@
 import { WrapPageElementBrowserArgs } from 'gatsby';
 import 'prismjs/themes/prism.css';
 import React from 'react';
-import Layout from './src/components/Layout';
+import { Default } from './src/layouts';
 import './src/styles/main.css';
 
 function wrapPageElement({ element, props }: WrapPageElementBrowserArgs): JSX.Element {
-  return <Layout {...props}>{element}</Layout>;
+  return <Default {...props}>{element}</Default>;
 }
 
 export { wrapPageElement };
