--- src/components/Layout/index.ts
@@ -1,4 +1,0 @@
-import Layout from './Layout';
-
-export default Layout;
-export * from './Layout';
