+++ src/components/index.ts
@@ -1,0 +1,26 @@
+export * from './Aside';
+export { default as Aside } from './Aside';
+export * from './Categories';
+export { default as Categories } from './Categories';
+export * from './Column';
+export { default as Column } from './Column';
+export * from './Content';
+export { default as Content } from './Content';
+export * from './Footer';
+export { default as Footer } from './Footer';
+export * from './GitHub';
+export { default as GitHub } from './GitHub';
+export * from './Grid';
+export { default as Grid } from './Grid';
+export * from './Header';
+export { default as Header } from './Header';
+export * from './Main';
+export { default as Main } from './Main';
+export * from './Navigation';
+export { default as Navigation } from './Navigation';
+export * from './Row';
+export { default as Row } from './Row';
+export * from './Section';
+export { default as Section } from './Section';
+export * from './SEO';
+export { default as SEO } from './SEO';
