+++ src/layouts/Default/index.ts
@@ -1,0 +1,4 @@
+import Default from './Default';
+
+export default Default;
+export * from './Default';
