+++ src/layouts/index.ts
@@ -1,0 +1,2 @@
+export * from './Default';
+export { default as Default } from './Default';
